scripts/pyproject.toml
[project]
name = "ocr-extractor"
version = "0.1.0"
description = "EasyOCR text extraction with structured data output"
requires-python = ">=3.8"
dependencies = [
    "easyocr>=1.7.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
